inv: merge via libtrx
