+++ src/libtrx/include/libtrx/game/inventory.h
@@ -1,10 +1,17 @@
 #pragma once
 
+#include "inventory_ring/types.h"
 #include "objects/ids.h"
 
 #include <stdbool.h>
 #include <stdint.h>
 
+bool Inv_AddItemNTimes(GAME_OBJECT_ID object_id, int32_t qty);
+GAME_OBJECT_ID Inv_GetItemOption(GAME_OBJECT_ID object_id);
+void Inv_InsertItem(INVENTORY_ITEM *inv_item);
+bool Inv_RemoveItem(GAME_OBJECT_ID object_id);
+int32_t Inv_RequestItem(GAME_OBJECT_ID object_id);
+void Inv_ClearSelection(void);
+void Inv_RemoveAllItems(void);
+
 extern bool Inv_AddItem(GAME_OBJECT_ID object_id);
-extern bool Inv_AddItemNTimes(GAME_OBJECT_ID object_id, int32_t qty);
-extern int32_t Inv_RequestItem(const GAME_OBJECT_ID object_id);
