+++ src/libtrx/include/libtrx/game/inventory_ring/vars.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "./types.h"
+
+extern INV_RING_SOURCE g_InvRing_Source[RT_NUMBER_OF];
