+++ src/tr1/game/inventory_ring/vars.h
@@ -3,10 +3,8 @@
 #include "global/types.h"
 
 #include <libtrx/game/clock.h>
-#include <libtrx/game/inventory_ring/types.h>
+#include <libtrx/game/inventory_ring/vars.h>
 #include <libtrx/game/text.h>
-
-extern INV_RING_SOURCE g_InvRing_Source[RT_NUMBER_OF];
 
 extern INVENTORY_ITEM g_InvRing_Item_Compass;
 extern INVENTORY_ITEM g_InvRing_Item_Medi;
