+++ src/tr2/game/inventory.h
@@ -1,11 +1,3 @@
 #pragma once
 
 #include <libtrx/game/inventory.h>
-#include <libtrx/game/inventory_ring/types.h>
-#include <libtrx/game/objects/types.h>
-
-void Inv_InsertItem(INVENTORY_ITEM *inv_item);
-int32_t Inv_RequestItem(GAME_OBJECT_ID object_id);
-void Inv_RemoveAllItems(void);
-bool Inv_RemoveItem(GAME_OBJECT_ID object_id);
-GAME_OBJECT_ID Inv_GetItemOption(GAME_OBJECT_ID object_id);
