+++ src/tr2/game/inventory_ring/control.h
@@ -9,5 +9,3 @@
 INV_RING *InvRing_Open(INVENTORY_MODE mode);
 GAME_FLOW_COMMAND InvRing_Control(INV_RING *ring, int32_t num_frames);
 GAME_FLOW_COMMAND InvRing_Close(INV_RING *ring);
-
-void InvRing_ClearSelection(void);
